requirements.txt
ffmpeg-python==0.2.0
matplotlib==3.8.2
numpy==1.26.2
scipy==1.11.4
